QNativeUIKitWindow: post updateLayout request when window becomes visible
When using QML, the window will be constructed first, and set to visible,
before constructing any of its children. So trying to update the layout
mmediatly upon the call to setVisible is pointless.

This patch will instead post the update request, and as such, let all the children
complete their initializtion before doing the layout.