Revert "Remove cpp/platforms/uikit/window example"
This reverts commit df0298d7da884755f268e94da5d9c422ceae2734.

Yes, keep this for now. I use it every day as the first example to
check if something goes wrong in UIKit. Both hybrid and native adds
extra complexity and logic to the mix. E.g if the uikit/window example
doesn't work, the other examples will for sure not work either. But not
necessarily the other way around.

We should go through the examples later and rename/remove them etc.