QNativePlatformBase: inherit from QNativeQbjectBase
A litte change to the structure. We now let QNativePlatformBase
inherit from QNativeObjectBase. Then we don't need to implement
the same QML logic on all the platforms, logic/dependency we
eventually would like to get rid of / change to make the cpp API
independent of QML. We also require the platform base classes to
be QObjects, which we now make more formal.