QNative: add QNativePlatformBase::setPlatformParent()
It turns out that using QObject::setParent directly has too many
issues, especially since it relies on the new parent to have
a private part that is finished contructing so that it can
inspect it in the QObject::childEvent handler. But this complicates
the platform plugins that may need to implement the platform
interfaces but without needing to use pimpl. Besides, it also
have QML parenting issues. So we 'revert' back to use
setPlatformParent that we have much better control over.