AppKit: don't lazy create viewcontroller views
AppKit might access NSViewController.view directly, in which
case a default view gets created if not yet set. And this view
will then be overwritten by QAppKitViewController::view(), the
way it stands. So we need to create it immediatly to ensure that
both Qt and AppKit ends up using the same view.