UIKit: don't lazy create viewcontroller views
UIKit might access UIViewController.view directly, in which
case a default view gets created if not yet set. And this view
will then be overwritten by QUIKitViewController::view(), the
way it stands. So we need to create it immediatly to ensure that
both Qt and UIKit ends up using the same view.