UIKit: remove bloated parent/child heuristics
It is probably vise to not overdo 'artifical' parent/child support when it doesn't
match the native API of doing things. It bloats the API by adding several ways of
achieving the same thing, it's prone to create problems if adding more than one control
as a child when jsut one is expected, and causes 'magic' to happen that can be hard to
support in the future. Having to write e.g 'tabBarItem: TabBarItem {}' in the QML code is
explicit and perfectly fine.