UIKit: use KVO for changes to a view frame
UIKit will sometimes change the frame of a view on it's own.
This will e.g happen for the root view inside a view controller
upon orientation change, or when the root view inside a tab is
made visible the first time.

Since we only wrap the native controls, and as such, cannot easily
override methods like "updateSubviews", we choose to use the
infamous KVO pattern to catch the frame changes for now, so that
we always emit signals when the frame changes.