UIKit: catch exception when trying to set an invalid view as a viewcontrollers view"
UIKit will throw an exception if trying to move a view from one view controller to
another without removing it from the first viewcontroller explicit. Since it is
not in the nature of QML to just crash for such issues, and since we at the same
time don't want to add to much logic to work around how UIKit works, we choose to
just catch the exception, and continue if possible.