UIKit: delay emitting unrelated geometry changed signals for QUniUIKitView
When setting a frame with a negative width/height on a UIView, UIKit
will modify the rect so that width/height end up positive, and
leave x/y negative instead. This causes a lot of trouble for
us, since during startup, both width and height bindings can
easily turn out negative until all properties are initialized.

And after a bit of study, it shows that we can only emit a changed
signal for the property being set. Otherwise, binding loops can easily
happen. So for that reason, we need to keep track of which properties
to emit when, to avoid hard-to-track binding loops to occur.