UIKit: change geometry singal emit strategy in QUniUIKitView
Now that we don't use KVO to detect geometry changes, we can remove
a lot of the structure implemented to handle synchronouse frame change
callbacks. Calling [view setNeedsLayout] is documented to always be
async, which means that we never end up in a situation where we emit
changes to several properties in one go after e.g just setting height.

We could in theory avoid emitting signals in the sync functions as well
since we sooner or later will get a layoutSubviews callback and emit
signals then. But we choose to be a bit pro-active and emit signals
early on to not undermine QMLs system for detecting dependency loops
(which we would if the signals were emitted async).